# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Productos",
        "description": "API REST para gestionar productos, categorías, etiquetas, pedidos y usuarios. Utiliza autenticación JWT: primero inicia sesión en POST /api/auth/login y luego envía el token en el header `Authorization: Bearer <token>`. El catálogo de productos es público; los pedidos y la gestión por rol exigen token.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://ense-ar-laravel.test",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/pedidos": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Listar pedidos",
                "description": "El cliente ve solo sus pedidos; el empleado y el admin ven todos.",
                "operationId": "e848822b988b4e4cf96af0f71f29581c",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de registros por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado paginado de pedidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Paginado"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido o ausente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Rol sin permisos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Crear un pedido",
                "description": "Crea un pedido con los productos indicados. El precio se toma como instantánea al momento de comprar.",
                "operationId": "36884e796fd6f470919017f7ca543cd0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items"
                                ],
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "producto_id",
                                                "cantidad"
                                            ],
                                            "properties": {
                                                "producto_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "cantidad": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "minItems": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string",
                                            "example": "Pedido creado con éxito"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PedidoResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido o ausente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Rol sin permisos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pedidos/{id}": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Mostrar un pedido",
                "description": "El cliente solo ve pedidos propios; el empleado y el admin ven cualquiera.",
                "operationId": "41f9fd012dcc6f1915be1a7372c7d87b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del pedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidoResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido o ausente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No puedes ver un pedido de otro usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pedido no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pedidos/{id}/estado": {
            "put": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Cambiar el estado de un pedido",
                "description": "Solo el admin puede mover los estados respetando las transiciones válidas.",
                "operationId": "f8a448d1ba08668fb9d8350859b4b12c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "estado"
                                ],
                                "properties": {
                                    "estado": {
                                        "type": "string",
                                        "example": "pagado",
                                        "enum": [
                                            "pendiente",
                                            "pagado",
                                            "enviado",
                                            "entregado",
                                            "cancelado"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estado actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string",
                                            "example": "Estado actualizado a Pagado"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PedidoResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido o ausente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Solo el admin",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pedido no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Transición de estado no válida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PedidoResource": {
                "required": [
                    "id",
                    "user_id",
                    "total",
                    "estado"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "total": {
                        "type": "number",
                        "format": "float",
                        "example": 2500
                    },
                    "estado": {
                        "type": "string",
                        "example": "pendiente",
                        "enum": [
                            "pendiente",
                            "pagado",
                            "enviado",
                            "entregado",
                            "cancelado"
                        ]
                    },
                    "usuario": {
                        "type": "string",
                        "example": "Juan Pérez",
                        "nullable": true
                    },
                    "productos": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "nombre": {
                                    "type": "string",
                                    "example": "Teclado Mecánico"
                                },
                                "cantidad": {
                                    "type": "integer",
                                    "example": 2
                                },
                                "precio": {
                                    "type": "number",
                                    "format": "float",
                                    "example": 1200.5
                                },
                                "subtotal": {
                                    "type": "number",
                                    "format": "float",
                                    "example": 2401
                                }
                            },
                            "type": "object"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "required": [
                    "error",
                    "message"
                ],
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "NOT_FOUND"
                    },
                    "message": {
                        "type": "string",
                        "example": "Recurso no encontrado"
                    },
                    "details": {
                        "description": "Detalles de errores de validación",
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Paginado": {
                "required": [
                    "estado",
                    "data"
                ],
                "properties": {
                    "estado": {
                        "type": "string",
                        "example": "exito"
                    },
                    "data": {
                        "required": [
                            "current_page",
                            "data",
                            "per_page",
                            "total"
                        ],
                        "properties": {
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "data": {
                                "description": "Elementos de la página",
                                "type": "array",
                                "items": {
                                    "type": "object"
                                }
                            },
                            "first_page_url": {
                                "type": "string",
                                "nullable": true
                            },
                            "from": {
                                "type": "integer",
                                "example": 1,
                                "nullable": true
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "last_page_url": {
                                "type": "string",
                                "nullable": true
                            },
                            "links": {
                                "type": "array",
                                "items": {
                                    "type": "object"
                                }
                            },
                            "next_page_url": {
                                "type": "string",
                                "nullable": true
                            },
                            "path": {
                                "type": "string",
                                "nullable": true
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 15
                            },
                            "prev_page_url": {
                                "type": "string",
                                "nullable": true
                            },
                            "to": {
                                "type": "integer",
                                "nullable": true
                            },
                            "total": {
                                "type": "integer",
                                "example": 42
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Token JWT obtenido en POST /api/auth/login. Enviar como: Authorization: Bearer <token>"
            }
        }
    },
...
}